Switch Tab component to use container component as base. Also re-arrange the tab related things to subfolder
